requirements.txt
azure-functions==1.23.0
azure-storage-blob==12.19.0
azure-storage-queue==12.9.0
Pillow>=10.2.0
requests==2.31.0



